chore: 调整geosite初始化位置
